+++ src/snabl/target.hpp
@@ -4,6 +4,7 @@
 #include "snabl/call.hpp"
 #include "snabl/ptrs.hpp"
 #include "snabl/state.hpp"
+#include "snabl/types.hpp"
 
 namespace snabl {
 	class Target {
@@ -11,8 +12,7 @@ namespace snabl {
 		enum class Opts: int {None=0, Vars=1, Recalls=2};
 		
 		Target(const ScopePtr &parent_scope=nullptr,
-					 ssize_t start_pc=-1,
-					 ssize_t end_pc=-1,
+					 Int start_pc=-1, Int end_pc=-1,
 					 Opts opts=Opts::None):
 			_parent_scope(parent_scope),
 			_start_pc(start_pc), _end_pc(end_pc),
@@ -20,11 +20,11 @@ namespace snabl {
 
 		virtual ~Target() { }
 		virtual string target_id() const=0;
-		ssize_t start_pc() const { return _start_pc; }
+		Int start_pc() const { return _start_pc; }
 		Opts opts() const { return _opts; }
 	protected:
 		ScopePtr _parent_scope;
-		ssize_t _start_pc, _end_pc;
+		Int _start_pc, _end_pc;
 		Opts _opts;
 	};
